+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * remove context from entryLogger
 * caution text to system YELLOW
 * adb print cleanup
+* mediaRecordPlacebo to VOICE_COMM source
 
 ### 4.4.0 (2020-6-27)
 * bugfix - add context for entryLogger