+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ### 4.4.1 (unreleased)
-* remove context from entryLogger
+* Android 10 concurrent audio mitigation test build
+* centered popup toast warning for concurrent audio capture state
+* passive jammer auto restart during C-1-4 conflict 
 * caution text to system YELLOW
-* adb print cleanup
 * mediaRecordPlacebo to VOICE_COMM source
 * browser intent ex null
+* add accessibility service permission to background check
+* updated Inspector and Readme text
 
 ### 4.4.0 (2020-6-27)
 * bugfix - add context for entryLogger